Update trend_tracking.sol
To begin with, both the users and the artists are required to either possess or newly create the wallets for NFTs upon their creation of a Spotify account. Once their wallet address is registered and tied to their Spotify address, the entire process described below can happen conveniently  in-app. 

For each brand-new song added to Spotify, a new NFT is created, with the initial ownership set to the wallet address of the artist. A user must first acquire a portion of ownership in a song and pay a royalty to the artist to be able to stream it. These operations are conducted by the functions in the smart contract, which are called upon simple user actions on the application interface such as clicking buttons. The transfer of ownership between users is disabled to preempt unpredictable impacts on the NFTs’ prices. When a song is streamed, its record is kept in the smart contract and can be used to create a music chart.